feat: move head & seo logic to default layout
to be able to use it in error
